requirements.txt
# Install with: conda install --file requirements.txt
flask==2.3.3
python-dotenv==1.0.0
openai==1.3.0
pymupdf==1.23.8
pypdf2==3.0.1
pandas==2.1.4
requests==2.31.0
werkzeug==2.3.7
pathlib==1.0.1
uuid==1.30

# python-dotenv==1.0.0 - For loading environment variables from .env file
# openai==1.3.0 - For SambaNova API integration
# PyMuPDF==1.23.8 - For advanced PDF handling (fitz)
# PyPDF2==3.0.1 - For PDF processing
# pandas==2.1.4 - For handling CSV and data structures
# requests==2.31.0 - For making HTTP requests to the API
# Flask==2.3.3 - For creating web applications
# Flask-WTF==1.1.1 - For handling web forms
# Werkzeug==2.3.7 - For handling WSGI requests
# pathlib==1.0.1 - For cross-platform path handling
# uuid==1.30 - For generating unique identifiers